Add files via upload
the Main Thread Demo program demonstrates the Main Thread and the use of some of its object methods.
When a Java program starts up, the main thread begins running immediately. 
The main thread is the thread from which other “child” threads will be spawned.
Often, it must be the last thread to finish execution because it performs various shutdown actions.

The second Main Thread program creates a thread by implementing Runnable interface and instantiating an object of type Thread. Runnable abstracts a unit of executable code. This class implement a single Runnable method called run(). Inside run() the code that constitutes the thread is defined. The run() method establishes the entry point for another, concurrent thread of execution within the  program.